Create README.md
I have not been in dev in over a decade. But, I will learn and do whatever I need to do to hopefully create a community-based and community-run learning environment that flourishes. I don't pretend to belong here, as 20 years in dev/qa/admin really doesn't mean much in such fast evolving field. But, I have been a teacher for 30 years in one way or another.